Implement Q-Learning?
I am unsure if this is implemented correctly. All the agent does right
now is fold. Given that the state size is so small, it may be the
optimal thing to do, but I am thinking we will need to introduce epsilon
into the picture so that it will explore more options rather than just
folding.

I am currently only storing the cards on the board and in the agents
hand as the state (as a frozenset) for the qvalue lookup table.  We
might have to design a custom intermediary data type for this that
hashes out as the sum of all the faces and suits of the cards (or
something similar). I believe that until we do, I cannot guarantee that
some states will errorneously be included more than once in the table.